msix: Added some checks and cleanup
